+++ source/public/Install-ChocolateySoftware.ps1
@@ -57,6 +57,10 @@
 .PARAMETER InstallationDirectory
     Set the Installation Directory for Chocolatey, by creating the Environment Variable. This will persist after the installation.
 
+.PARAMETER RunNonElevated
+    Throws if the process is not running elevated. use -RunNonElevated if you really want to run
+    even if the current shell is not elevated.
+
 .EXAMPLE
     Install latest chocolatey software from the Community repository (non pre-release version)
     Install-ChocolateySoftware
@@ -70,6 +74,7 @@
 #>
 function Install-ChocolateySoftware
 {
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSReviewUnusedParameter', '')]
     [CmdletBinding(
         DefaultParameterSetName = 'FromFeedUrl'
     )]
@@ -108,7 +113,11 @@ function Install-ChocolateySoftware
 
         [Parameter()]
         [System.String]
-        $InstallationDirectory
+        $InstallationDirectory,
+
+        [Parameter(DontShow)]
+        [switch]
+        $RunNonElevated = $(Assert-ChocolateyIsElevated)
     )
 
     if ($PSVersionTable.PSVersion.Major -lt 4)