+++ source/public/Uninstall-ChocolateyPackage.ps1
@@ -109,6 +109,10 @@
     uninstaller reports an error. Overrides the default feature
     'failOnAutoUninstaller' set to 'False'. Available in 0.9.10+.
 
+.PARAMETER RunNonElevated
+    Throws if the process is not running elevated. use -RunNonElevated if you really want to run
+    even if the current shell is not elevated.
+
 .EXAMPLE
     Uninstall-ChocolateyPackage -Name Putty
 
@@ -117,6 +121,7 @@
 #>
 function Uninstall-ChocolateyPackage
 {
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSReviewUnusedParameter', '')]
     [CmdletBinding(
         SupportsShouldProcess = $true,
         ConfirmImpact = 'High'
@@ -260,7 +265,11 @@ function Uninstall-ChocolateyPackage
             ValueFromPipelineByPropertyName
         )]
         [Switch]
-        $IgnoreAutoUninstallerFailure
+        $IgnoreAutoUninstallerFailure,
+
+        [Parameter(DontShow)]
+        [switch]
+        $RunNonElevated = $(Assert-ChocolateyIsElevated)
     )
 
     begin