+++ source/public/Unregister-ChocolateySource.ps1
@@ -50,6 +50,10 @@
     Do Not Show Progress - Do not show download progress percentages.
     Available in 0.10.4+.
 
+.PARAMETER RunNonElevated
+    Throws if the process is not running elevated. use -RunNonElevated if you really want to run
+    even if the current shell is not elevated.
+
 .EXAMPLE
     Unregister-ChocolateySource -Name MyProgetFeed
 
@@ -58,6 +62,7 @@
 #>
 function Unregister-ChocolateySource
 {
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSReviewUnusedParameter', '')]
     [CmdletBinding()]
     param (
         [Parameter(
@@ -118,8 +123,11 @@ function Unregister-ChocolateySource
             ValueFromPipelineByPropertyName
         )]
         [Switch]
-        $NoProgress
+        $NoProgress,
 
+        [Parameter(DontShow)]
+        [switch]
+        $RunNonElevated = $(Assert-ChocolateyIsElevated)
     )
 
     process